+++ projects/design-angular-kit/src/lib/components/core/avatar/avatar-group/avatar-group.component.ts
@@ -1,11 +1,26 @@
-import { Component, ContentChildren, HostBinding, Input, QueryList, TemplateRef, ViewChild } from '@angular/core';
 import { NgTemplateOutlet } from '@angular/common';
+import { Component, ContentChildren, Directive, HostBinding, Input, QueryList, TemplateRef, ViewChild } from '@angular/core';
 import { inputToBoolean } from '../../../../utils/coercion';
+
+@Directive({
+  standalone: true,
+  selector: '[itAvatarListItem]',
+  exportAs: 'itAvatarListItem',
+})
+export class ItAvatarListItemDirective {
+  @HostBinding('class')
+  get hostClasses(): string {
+    const cssClass = `list-item`;
+    return cssClass;
+  }
+}
 
 @Component({
   standalone: true,
   selector: 'it-avatar-item',
-  template: '<ng-template><ng-content></ng-content></ng-template>',
+  template: `<ng-template let-linkListItem="linkList">
+    <ng-content></ng-content>
+  </ng-template>`,
 })
 export class ItAvatarGroupItemComponent {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -14,8 +29,16 @@ export class ItAvatarGroupItemComponent {
 
 @Component({
   selector: 'it-avatar-group',
-  templateUrl: './avatar-group.component.html',
   imports: [NgTemplateOutlet],
+  template: `
+    <ul [class]="linkList ? 'link-list avatar-group' : 'avatar-group-stacked'">
+      @for (avatar of avatars; track avatar) {
+        <li>
+          <ng-container *ngTemplateOutlet="avatar._implicitContent; context: { linkList: linkList }"></ng-container>
+        </li>
+      }
+    </ul>
+  `,
 })
 export class ItAvatarGroupComponent {
   @Input({ transform: inputToBoolean }) @HostBinding('class.link-list-wrapper') linkList = false;