+++ src/app/autocomplete/autocomplete-big-search-example/autocomplete-big-search-example.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-import { AutocompleteSearchExampleComponent } from '../autocomplete-search-example/autocomplete-search-example.component';
-
-@Component({
-  selector: 'it-autocomplete-big-search-example',
-  templateUrl: './autocomplete-big-search-example.component.html',
-  standalone: false,
-})
-export class AutocompleteBigSearchExampleComponent extends AutocompleteSearchExampleComponent {}