+++ src/app/autocomplete/autocomplete-examples/autocomplete-examples.component.spec.ts
@@ -1,4 +1,4 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { AutocompleteExamplesComponent } from './autocomplete-examples.component';
 
@@ -6,10 +6,13 @@ describe('AutocompleteExamplesComponent', () => {
   let component: AutocompleteExamplesComponent;
   let fixture: ComponentFixture<AutocompleteExamplesComponent>;
 
-  beforeEach(() => {
+  beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [AutocompleteExamplesComponent],
-    });
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
     fixture = TestBed.createComponent(AutocompleteExamplesComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();