+++ src/app/autocomplete/autocomplete-index/autocomplete-index.component.spec.ts
@@ -1,4 +1,4 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { AutocompleteIndexComponent } from './autocomplete-index.component';
 
@@ -6,10 +6,13 @@ describe('AutocompleteIndexComponent', () => {
   let component: AutocompleteIndexComponent;
   let fixture: ComponentFixture<AutocompleteIndexComponent>;
 
-  beforeEach(() => {
+  beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [AutocompleteIndexComponent],
-    });
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
     fixture = TestBed.createComponent(AutocompleteIndexComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();