+++ src/app/autocomplete/autocomplete-routing.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { Routes, RouterModule } from '@angular/router';
 import { AutocompleteIndexComponent } from './autocomplete-index/autocomplete-index.component';
 
 const routes: Routes = [{ path: '', component: AutocompleteIndexComponent }];