+++ src/app/autocomplete/autocomplete.module.ts
@@ -2,21 +2,16 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { AutocompleteRoutingModule } from './autocomplete-routing.module';
-import { AutocompleteIndexComponent } from './autocomplete-index/autocomplete-index.component';
-import { ItTabContainerComponent } from 'design-angular-kit/components/core/tab/tab-container/tab-container.component';
-import { ItTabItemComponent } from 'design-angular-kit/components/core/tab/tab-item/tab-item.component';
-import { SharedModule } from '../shared/shared.module';
+import { AutocompleteExampleComponent } from './autocomplete-example/autocomplete-example.component';
+import { AutocompleteDoubleComponent } from './autocomplete-double/autocomplete-double.component';
+
 import { AutocompleteExamplesComponent } from './autocomplete-examples/autocomplete-examples.component';
-import { AutocompleteSearchExampleComponent } from './autocomplete-search-example/autocomplete-search-example.component';
-import { AutocompleteBigSearchExampleComponent } from './autocomplete-big-search-example/autocomplete-big-search-example.component';
+import { AutocompleteIndexComponent } from './autocomplete-index/autocomplete-index.component';
+import { SharedModule } from '../shared/shared.module';
+import { FormsModule } from '@angular/forms';
 
 @NgModule({
-  declarations: [
-    AutocompleteIndexComponent,
-    AutocompleteExamplesComponent,
-    AutocompleteSearchExampleComponent,
-    AutocompleteBigSearchExampleComponent,
-  ],
-  imports: [CommonModule, AutocompleteRoutingModule, ItTabContainerComponent, ItTabItemComponent, SharedModule],
+  imports: [CommonModule, FormsModule, SharedModule, AutocompleteRoutingModule],
+  declarations: [AutocompleteIndexComponent, AutocompleteExampleComponent, AutocompleteExamplesComponent, AutocompleteDoubleComponent],
 })
 export class AutocompleteModule {}