+++ src/app/footer/footer-routing.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { FooterIndexComponent } from './footer-index/footer-index.component';
+
+const routes: Routes = [{ path: '', component: FooterIndexComponent }];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class FooterRoutingModule {}