+++ src/app/search/search-examples/search-examples.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'it-search-examples',
+  templateUrl: './search-examples.component.html',
+})
+export class SearchExamplesComponent {}