+++ src/main.ts
@@ -1,4 +1,4 @@
-import { enableProdMode } from '@angular/core';
+import { ApplicationRef, enableProdMode, isDevMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
@@ -8,6 +8,14 @@ if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic()
-  .bootstrapModule(AppModule)
-  .catch(err => console.log(err));
+async function bootstrap() {
+  const { injector } = await platformBrowserDynamic().bootstrapModule(AppModule);
+
+  if (isDevMode()) {
+    const { enableDebugTools } = await import('@angular/platform-browser');
+    const { components } = injector.get(ApplicationRef);
+    enableDebugTools(components[0]);
+  }
+}
+
+bootstrap();